requirements.txt
click==8.1.3
joblib==1.3.2
numpy==1.26.3
pandas==2.1.4
python-dotenv==1.0.0
PyYAML==6.0.1
scikit-learn==1.3.2
setuptools==68.2.2
